Split RestartIO into Separate Load/Save Files
This is in preparation of introducing support for loading most
well-related solution items from the regular vectors IWEL, XWEL,
ICON, and XCON.
